user_levelsのコントローラを削除 #68
