--- spec/requests/api/user_levels_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "UserLevels", type: :request do
-
-end
